In NQueen.GUI and NQueen.Console:   - Changed all occurances of BoardSize type from byte to int

+++ NQueen.GUI/ViewModels/MainViewModel.Methods.cs
@@ -135,7 +135,7 @@ public sealed partial class MainViewModel
         }
     }
 
-    partial void OnBoardSizeChanged(byte value)
+    partial void OnBoardSizeChanged(int value)
     {
         IsValid = InputViewModel.Validate(this).IsValid;
